+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.4.5] - ???
+### Added
+- Support `-di` flag for `html5-list` command ([#43](https://github.com/SAP/cf-html5-apps-repo-cli-plugin/issues/43))
+- Support `-di` flag for `html5-push` command ([#44](https://github.com/SAP/cf-html5-apps-repo-cli-plugin/issues/44))
+- Support `-di` flag for `html5-delete` command ([#45](https://github.com/SAP/cf-html5-apps-repo-cli-plugin/issues/45))
+
+### Fixed
+- Requests to Destination configuration service now have the correct `Content-Type` header value (`application/json`)
+- Name of `app-host` service instance in case `sap.cloud/service` is not set in `manifest.json`
+
 ## [1.4.4] - 2020-10-19
 ### Added
 - Support new business service destination configuration format in `html5-list -d` and